Revert "Aula 4 (#2)"
This reverts commit 082af44612b205f79d1a2681b9e66d190f41b47e.
