--- .firebaserc
@@ -1,5 +1,0 @@
-{
-  "projects": {
-    "default": "letmeask-77895"
-  }
-}
