--- database.rules.json
@@ -1,20 +1,0 @@
-{
-  "rules": {
-    "rooms": {
-      ".read": false,
-      ".write": "auth != null",
-      "$roomId": {
-        ".read": true,
-        ".write": "auth != null && (!data.exists() || data.child('authorId').val() == auth.id)",
-        "questions": {
-          ".read": true,
-          ".write": "auth != null && (!data.exists() || data.parent().child('authorId').val() == auth.id)",
-          "likes": {
-            ".read": true,
-            ".write": "auth != null && (!data.exists() || data.parent().child('authorId').val() == auth.id)"
-          }
-        }
-      }
-    }
-  }
-}
