+++ package.json
@@ -10,7 +10,6 @@
     "@types/node": "^12.0.0",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
-    "classnames": "^2.3.1",
     "firebase": "^8.6.8",
     "node-sass": "5.0.0",
     "react": "^17.0.2",
