+++ src/App.tsx
@@ -4,7 +4,6 @@ import { AuthContextProvider } from "./contexts/AuthContext"
 import { Home } from "./pages/Home"
 import { NewRoom } from "./pages/NewRoom"
 import { Room } from './pages/Room'
-import { AdminRoom } from './pages/AdminRoom'
 
 function App() {
   return (
@@ -14,7 +13,6 @@ function App() {
           <Route path="/" exact component={ Home } />
           <Route path="/rooms/new" component={ NewRoom } />
           <Route path="/rooms/:id" component={ Room } />
-          <Route path="/admin/rooms/:id" component={ AdminRoom } />
         </Switch>
       </AuthContextProvider>
     </BrowserRouter>
