+++ src/components/Button.tsx
@@ -2,12 +2,10 @@ import { ButtonHTMLAttributes } from 'react'
 
 import '../styles/button.scss'
 
-type ButtonProps = ButtonHTMLAttributes<HTMLButtonElement> & { 
-    isOutlined?: boolean
-}
+type ButtonProps = ButtonHTMLAttributes<HTMLButtonElement>
 
-export function Button({ isOutlined = false, ...props }: ButtonProps) {
+export function Button(props: ButtonProps) {
     return (
-        <button className={`button ${isOutlined ? 'outlined' : ''}`} { ...props }></button>
+        <button className="button" { ...props }></button>
     )
 }
