+++ src/styles/button.scss
@@ -18,12 +18,6 @@
         margin-right: 8px;
     }
 
-    &.outlined {
-        background-color: #FFFFFF;
-        border: 1px solid #835AFD;
-        color: #835AFD;
-    }
-
     &:not(:disabled):hover {
         filter: brightness(0.9);
     }
