+++ src/styles/room.scss
@@ -13,15 +13,6 @@
             > img {
                 max-height: 45px;
             }
-
-            > div {
-                display: flex;
-                gap: 16px;
-
-                button {
-                    height: 40px;
-                }
-            }
         }
     }
 
@@ -102,10 +93,6 @@
                     }
                 }
             }
-        }
-
-        .question-list {
-            margin-top: 32px;
         }
     }
 }
